Add a variable at the top of the Makefile to set the log level.
# Aider chat conversation:

USER: How is log level set in the makefile?

ASSISTANT: The log level in the `Makefile` is set using an environment variable `LOG_LEVEL`. This environment variable is checked within the `Makefile` to determine if additional debug information should be displayed. The `Makefile` does not set the `LOG_LEVEL` itself; it is expected to be set externally by the user or the environment in which the `Makefile` is run.

For example, you can set the log level to debug in your shell before running the `make` command like this:

```sh
export LOG_LEVEL=DEBUG
make push
```

Or you can set it inline with the `make` command:

```sh
LOG_LEVEL=DEBUG make push
```

The `Makefile` then checks for this variable with conditional statements. If `LOG_LEVEL` is set to `DEBUG`, it will execute additional `echo` commands to display the debug information.

USER: Lets add a variable at top of the makefile

ASSISTANT: To add a variable at the top of the `Makefile` for setting the log level, I will introduce a new line that defines a `LOG_LEVEL` variable. This variable can then be used throughout the `Makefile` to control the verbosity of the logging.

Here is the diff for the `Makefile`:

```diff
--- Makefile
+++ Makefile
@@ ... @@
+LOG_LEVEL ?= INFO
+
 USERNAME := $(shell whoami)
 TIMESTAMP := $(shell date +%Y%m%d%H%M%S)
 BRANCH_NAME := $(USERNAME)-$(TIMESTAMP)
```

This change introduces a `LOG_LEVEL` variable with a default value of `INFO`. The `?=` operator is used to set the variable only if it is not already set in the environment, allowing you to override it by exporting `LOG_LEVEL` with a different value before running `make`.