Remove duplicate `push:` target label in the `Makefile`.
# Aider chat conversation:

USER: I ran this command:

 make push

And got this output:

/bin/sh: push:: command not found
make: *** [push] Error 127

ASSISTANT: It seems there is a duplicate `push:` target label in the `Makefile` which is causing the error. I will provide a diff to fix this issue by removing the duplicate label.

```diff
--- Makefile
+++ Makefile
@@ ... @@
 branch:
 	@$(MAKE) confirm_branch

-confirm_branch:
+confirm_branch:
@@ ... @@
-push:
-	@current_branch=$$(git rev-parse --abbrev-ref HEAD); \
 push:
 	@set -e; \
 	current_branch=$$(git rev-parse --abbrev-ref HEAD); \
```

Apply this patch to the `Makefile` to correct the syntax error.